requirements.txt
cryptography==3.4.7
numpy==1.19.5
greenlet==0.4.15
simplejson==3.17.0
lxml==4.6.2
dataclasses==0.8
contextvars==2.4
Babel==2.9.0
blinker==1.4
curio==1.5
Cython==0.29.23
flask_babel==2.0.0
flask_babelex==0.9.4
ipaddr==2.2.0
ordereddict==1.1
pretty==0.1
protobuf==3.15.8
pyOpenSSL==20.0.1
python-dotenv==0.17.0
requests_toolbelt==0.9.1
sniffio==1.2.0
trio==0.18.0
watchdog==2.0.2
wincertstore==0.2
flask-wtf==0.14.3
Authlib==0.15.3
